doneee

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: djin <djin@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/02 14:15:17 by djin              #+#    #+#             */
-/*   Updated: 2023/06/06 01:01:22 by djin             ###   ########.fr       */
+/*   Updated: 2023/06/06 15:41:45 by djin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,11 +24,11 @@
 
 int		ft_printf(const char *str, ...);
 void	print_ptr(unsigned long int n, int *len);
-void	print_hex(unsigned long int n, int *len);
-void	print_hex_upper(unsigned long int n, int *len);
+void	print_hex(unsigned int n, int *len);
+void	print_hex_upper(unsigned int n, int *len);
 void	ft_put_char(char c, int *len);
 void	ft_put_str(char *str, int *len);
-void	print_unsigned(unsigned long n, int *len);
+void	print_unsigned(unsigned int n, int *len);
 void	put_nbr(long n, int *len);
 
 #endif
